annotators/midas_predictor/requirements.txt
Flask==1.1.1
scikit-learn==1.0.1
typing==3.7.4.1
gunicorn==19.9.0
requests==2.22.0
sentry-sdk[asgi]==1.3.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-hub==0.12.0
itsdangerous==2.0.1
pickle5==0.0.12
